Refactor concrete LocalFileSystem into own namespace
