+++ src/Liquid/LocalFileSystem.php
@@ -12,88 +12,6 @@
 namespace Liquid;
 
 /**
- * This implements an abstract file system which retrieves template files named in a manner similar to Rails partials,
- * ie. with the template name prefixed with an underscore. The extension ".liquid" is also added.
- *
- * For security reasons, template paths are only allowed to contain letters, numbers, and underscore.
+ * @deprecated Left for backward compatibility reasons. Use \Liquid\FileSystem\Local instead.
  */
-class LocalFileSystem implements FileSystem
-{
-	/**
-	 * The root path
-	 *
-	 * @var string
-	 */
-	private $root;
-
-	/**
-	 * Constructor
-	 *
-	 * @param string $root The root path for templates
-	 */
-	public function __construct($root) {
-		// since root path can only be set from constructor, we check it once right here
-		if (!empty($root)) {
-			$realRoot = realpath($root);
-			if ($realRoot === false) {
-				throw new LiquidException("Root path could not be found: '$root'");
-			}
-			$root = $realRoot;
-		}
-
-		$this->root = $root;
-	}
-
-	/**
-	 * Retrieve a template file
-	 *
-	 * @param string $templatePath
-	 *
-	 * @return string template content
-	 */
-	public function readTemplateFile($templatePath) {
-		return file_get_contents($this->fullPath($templatePath));
-	}
-
-	/**
-	 * Resolves a given path to a full template file path, making sure it's valid
-	 *
-	 * @param string $templatePath
-	 *
-	 * @throws LiquidException
-	 * @return string
-	 */
-	public function fullPath($templatePath) {
-		if (empty($templatePath)) {
-			throw new LiquidException("Empty template name");
-		}
-
-		$nameRegex = Liquid::get('INCLUDE_ALLOW_EXT')
-			? new Regexp('/^[^.\/][a-zA-Z0-9_\.\/]+$/')
-			: new Regexp('/^[^.\/][a-zA-Z0-9_\/]+$/');
-
-		if (!$nameRegex->match($templatePath)) {
-			throw new LiquidException("Illegal template name '$templatePath'");
-		}
-
-		$templateDir = dirname($templatePath);
-		$templateFile = basename($templatePath);
-
-		if (!Liquid::get('INCLUDE_ALLOW_EXT')) {
-			$templateFile = Liquid::get('INCLUDE_PREFIX') . $templateFile . '.' . Liquid::get('INCLUDE_SUFFIX');
-		}
-
-		$fullPath = join(DIRECTORY_SEPARATOR, array($this->root, $templateDir, $templateFile));
-
-		$realFullPath = realpath($fullPath);
-		if ($realFullPath === false) {
-			throw new LiquidException("File not found: $fullPath");
-		}
-
-		if (strpos($realFullPath, $this->root) !== 0) {
-			throw new LiquidException("Illegal template full path: {$realFullPath} not under {$this->root}");
-		}
-
-		return $realFullPath;
-	}
-}
+class LocalFileSystem extends \Liquid\FileSystem\Local {}
